requirement.txt
tensorflow==2.6.0
numpy==1.19.5
pandas==1.3.4
matplotlib==3.4.3
opencv-python==4.5.4.58
Pillow==8.4.0
